renamed HomeView.vue to BrowseView.vue
